# open-api/aspnetcore/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "contact" : {
      "email" : "apiteam@swagger.io"
    },
    "description" : "This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about\nSwagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!\nYou can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features in OAS3.\n\n_If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_\n\nSome useful links:\n- [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)\n- [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService" : "http://swagger.io/terms/",
    "title" : "Swagger Petstore - OpenAPI 3.0",
    "version" : "1.0.11"
  },
...
  "servers" : [ {
    "url" : "https://petstore3.swagger.io/api/v3"
  } ],
...
  "paths" : {
    "/pet" : {
      "post" : {
        "description" : "Add a new pet to the store",
        "operationId" : "addPet",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Pet"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/Pet"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/Pet"
              }
            }
          },
          "description" : "Create a new pet in the store",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pet"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pet"
                }
              }
            },
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          }
        },
        "security" : [ {
          "petstore_auth" : [ "write:pets", "read:pets" ]
        } ],
        "summary" : "Add a new pet to the store",
        "tags" : [ "pet" ]
      },
      "put" : {
        "description" : "Update an existing pet by Id",
        "operationId" : "updatePet",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Pet"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/Pet"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/Pet"
              }
            }
          },
          "description" : "Update an existent pet in the store",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pet"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pet"
                }
              }
            },
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Pet not found"
          },
          "422" : {
            "description" : "Validation exception"
          }
        },
        "security" : [ {
          "petstore_auth" : [ "write:pets", "read:pets" ]
        } ],
        "summary" : "Update an existing pet",
        "tags" : [ "pet" ]
      }
    },
    "/pet/findByStatus" : {
      "get" : {
        "description" : "Multiple status values can be provided with comma separated strings",
        "operationId" : "findPetsByStatus",
        "parameters" : [ {
          "description" : "Status values that need to be considered for filter",
          "explode" : true,
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "default" : "available",
            "enum" : [ "available", "pending", "sold" ],
            "type" : "string"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Pet"
                  },
                  "type" : "array"
                }
              },
              "application/xml" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Pet"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid status value"
          }
        },
        "security" : [ {
          "petstore_auth" : [ "write:pets", "read:pets" ]
        } ],
        "summary" : "Finds Pets by status",
        "tags" : [ "pet" ]
      }
    },
    "/pet/findByTags" : {
      "get" : {
        "description" : "Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.",
        "operationId" : "findPetsByTags",
        "parameters" : [ {
          "description" : "Tags to filter by",
          "explode" : true,
          "in" : "query",
          "name" : "tags",
          "required" : false,
          "schema" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Pet"
                  },
                  "type" : "array"
                }
              },
              "application/xml" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Pet"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid tag value"
          }
        },
        "security" : [ {
          "petstore_auth" : [ "write:pets", "read:pets" ]
        } ],
        "summary" : "Finds Pets by tags",
        "tags" : [ "pet" ]
      }
    },
    "/pet/{petId}" : {
      "delete" : {
        "description" : "delete a pet",
        "operationId" : "deletePet",
        "parameters" : [ {
          "description" : "",
          "explode" : false,
          "in" : "header",
          "name" : "api_key",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "description" : "Pet id to delete",
          "explode" : false,
          "in" : "path",
          "name" : "petId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid pet value"
          }
        },
        "security" : [ {
          "petstore_auth" : [ "write:pets", "read:pets" ]
        } ],
        "summary" : "Deletes a pet",
        "tags" : [ "pet" ]
      },
      "get" : {
        "description" : "Returns a single pet",
        "operationId" : "getPetById",
        "parameters" : [ {
          "description" : "ID of pet to return",
          "explode" : false,
          "in" : "path",
          "name" : "petId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pet"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Pet"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Pet not found"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        }, {
          "petstore_auth" : [ "write:pets", "read:pets" ]
        } ],
        "summary" : "Find pet by ID",
        "tags" : [ "pet" ]
      },
      "post" : {
        "description" : "",
        "operationId" : "updatePetWithForm",
        "parameters" : [ {
          "description" : "ID of pet that needs to be updated",
          "explode" : false,
          "in" : "path",
          "name" : "petId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        }, {
          "description" : "Name of pet that needs to be updated",
          "explode" : true,
          "in" : "query",
          "name" : "name",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "Status of pet that needs to be updated",
          "explode" : true,
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "petstore_auth" : [ "write:pets", "read:pets" ]
        } ],
        "summary" : "Updates a pet in the store with form data",
        "tags" : [ "pet" ]
      }
    },
    "/pet/{petId}/uploadImage" : {
      "post" : {
        "description" : "",
        "operationId" : "uploadFile",
        "parameters" : [ {
          "description" : "ID of pet to update",
          "explode" : false,
          "in" : "path",
          "name" : "petId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        }, {
          "description" : "Additional Metadata",
          "explode" : true,
          "in" : "query",
          "name" : "additionalMetadata",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        } ],
        "requestBody" : {
          "content" : {
            "application/octet-stream" : {
              "schema" : {
                "format" : "binary",
                "type" : "string"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            },
            "description" : "successful operation"
          }
        },
        "security" : [ {
          "petstore_auth" : [ "write:pets", "read:pets" ]
        } ],
        "summary" : "uploads an image",
        "tags" : [ "pet" ]
      }
    },
...
  },
  "components" : {
...
    "schemas" : {
...
      "Category" : {
        "example" : {
          "name" : "Dogs",
          "id" : 1
        },
        "properties" : {
          "id" : {
            "example" : 1,
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "example" : "Dogs",
            "type" : "string"
          }
        },
        "type" : "object",
        "xml" : {
          "name" : "category"
        }
      },
...
      "Tag" : {
        "example" : {
          "name" : "name",
          "id" : 0
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object",
        "xml" : {
          "name" : "tag"
        }
      },
      "Pet" : {
        "example" : {
          "photoUrls" : [ "photoUrls", "photoUrls" ],
          "name" : "doggie",
          "id" : 10,
          "category" : {
            "name" : "Dogs",
            "id" : 1
          },
          "tags" : [ {
            "name" : "name",
            "id" : 0
          }, {
            "name" : "name",
            "id" : 0
          } ],
          "status" : "available"
        },
        "properties" : {
          "id" : {
            "example" : 10,
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "example" : "doggie",
            "type" : "string"
          },
          "category" : {
            "$ref" : "#/components/schemas/Category"
          },
          "photoUrls" : {
            "items" : {
              "type" : "string",
              "xml" : {
                "name" : "photoUrl"
              }
            },
            "type" : "array",
            "xml" : {
              "wrapped" : true
            }
          },
          "tags" : {
            "items" : {
              "$ref" : "#/components/schemas/Tag"
            },
            "type" : "array",
            "xml" : {
              "wrapped" : true
            }
          },
          "status" : {
            "description" : "pet status in the store",
            "enum" : [ "available", "pending", "sold" ],
            "type" : "string"
          }
        },
        "required" : [ "name", "photoUrls" ],
        "type" : "object",
        "xml" : {
          "name" : "pet"
        }
      },
      "ApiResponse" : {
        "example" : {
          "code" : 0,
          "type" : "type",
          "message" : "message"
        },
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "type" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object",
        "xml" : {
          "name" : "##default"
        }
      }
    },
    "securitySchemes" : {
      "petstore_auth" : {
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://petstore3.swagger.io/oauth/authorize",
            "scopes" : {
              "write:pets" : "modify pets in your account",
              "read:pets" : "read your pets"
            }
          }
        },
        "type" : "oauth2"
      },
      "api_key" : {
        "in" : "header",
        "name" : "api_key",
        "type" : "apiKey"
      }
    }
  }
}